Добавлен модуль data, в который перенесены api, datasource и реализация репозиториев. В модуле domain оставлен usecases и интерфейс репозитория. Все настройки по доступу к апи перенесены из domain (build.gradle.kts) в data (build.gradle.kts) UseCase модуля Domain теперь общаются в модулем Data посредством интерфейса репозитория и не знают о реализации репозитория и все что с ним связано. А вьюмодели работают только с UseCase. Вроде достигнута какая-то распределенность. Рефакторинг.
