+++ app/src/main/java/com/mwaiterdev/waiter/di/Di.kt
@@ -3,14 +3,14 @@ package com.mwaiterdev.waiter.di
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
-import com.mwaiterdev.domain.BuildConfig
-import com.mwaiterdev.domain.api.WaiterApi
-import com.mwaiterdev.domain.api.WaiterApiInterceptor
+import com.mwaiterdev.data.BuildConfig
+import com.mwaiterdev.data.api.WaiterApi
+import com.mwaiterdev.data.api.WaiterApiInterceptor
+import com.mwaiterdev.data.repository.RepositoryImpl
+import com.mwaiterdev.data.repository.RepositoryMockImp
+import com.mwaiterdev.data.repository.datasource.IRemoteDataSource
+import com.mwaiterdev.data.repository.datasource.RemoteDataSourceImpl
 import com.mwaiterdev.domain.repository.Repository
-import com.mwaiterdev.domain.repository.RepositoryImpl
-import com.mwaiterdev.domain.repository.RepositoryMockImp
-import com.mwaiterdev.domain.repository.datasource.IRemoteDataSource
-import com.mwaiterdev.domain.repository.datasource.RemoteDataSourceImpl
 import com.mwaiterdev.domain.usecase.billscreen.BillInteractorImpl
 import com.mwaiterdev.domain.usecase.billscreen.IBillInteractor
 import com.mwaiterdev.domain.usecase.mainbillsscreen.MainBillsIteractor
