+++ build.gradle.kts
@@ -7,7 +7,6 @@ buildscript {
     dependencies {
         classpath("com.android.tools.build:gradle:${LibVersion.GRADLE_VERSION}")
         classpath("org.jetbrains.kotlin:${LibVersion.GRADLE_PLUGIN_VERSION}")
-        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.0")
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
