+++ domain/build.gradle.kts
@@ -1,43 +1,17 @@
-import java.io.FileInputStream
-import java.util.*
-
 plugins {
-    id("com.android.library")
-    id("kotlin-parcelize")
-    kotlin("android")
-    kotlin("kapt")
+    id("java-library")
+    id("kotlin")
 }
 
-android {
-    compileSdk = Config.COMPILE_SDK
-
-    defaultConfig {
-        minSdk = Config.MIN_SDK_VERSION
-        targetSdk = Config.TARGET_SDK
-    }
-
-    buildTypes.forEach {
-        val properties = Properties()
-        properties.load(FileInputStream(file("./../conf.properties")))
-        val urlBase = properties.getProperty("base_url", "")
-        it.buildConfigField("String", "BASE_URL", urlBase)
-        val appId = properties.getProperty("appid", "")
-        it.buildConfigField("String", "APP_ID", appId)
-        val appToken = properties.getProperty("token", "")
-        it.buildConfigField("String", "API_TOKEN", appToken)
-    }
+java {
+    sourceCompatibility = Config.java_version
+    targetCompatibility = Config.java_version
 }
 
 dependencies {
 
     // Kotlin
-    implementation(Kotlin.CORE)
     implementation(Kotlin.STDLIB)
-
-    // Design
-    implementation(Design.APPCOMPAT)
-    implementation(Design.MATERIAL)
-    implementation(Design.CONSTRAINT_LAYOUT)
 
     // Retrofit
     implementation(Retrofit2.RETROFIT)
@@ -47,6 +21,4 @@ dependencies {
 
     //Tests
     testImplementation(Tests.JUNIT)
-    androidTestImplementation(Tests.TEST_EXT_JUNIT)
-    androidTestImplementation(Tests.ESPRESSO)
 }
