+++ settings.gradle
@@ -3,3 +3,4 @@ include ':app'
 include ':utils'
 include ':core'
 include ':domain'
+include ':data'
